Add refundable maximum
